+++ index.html
@@ -236,7 +236,6 @@
 
 </body>
 
-<!-- <script src="js/functions.js" type="module"></script> -->
 <script src="js/main.js" type="module"></script>
 
 </html>
